created practically everything needed. Now to tweek things and add passport auth sometime

+++ src/routes/room/deleteRoom.ts
@@ -4,10 +4,10 @@ import RoomModel from '../../models/room'
 const deleteRoom = express.Router()
 
 deleteRoom.delete('/', authUser, async (req, res) => {
-  const { _id } = req.body
-  
+  const { _id, userAccessId } = req.body
+
   try {
-    await RoomModel.deleteOne({ _id })
+    await RoomModel.deleteOne({ _id, users: userAccessId })
     res.status(200).send('deleted room')
   } catch (error) {
     console.log(error)
